Add data type check for BN weight/bias

+++ tests/gpu/examples/simple/test_lazy_reorder_batchnorm.py
@@ -6,7 +6,7 @@ import copy
 conv = nn.Conv2d(64, 64, kernel_size=3, stride=1, padding=1, bias=False)
 conv_xpu = copy.deepcopy(conv).bfloat16().to("xpu")
 bn = nn.BatchNorm2d(64)
-bn_xpu = copy.deepcopy(bn).bfloat16().to("xpu")
+bn_xpu = copy.deepcopy(bn).to("xpu")
 
 
 input = torch.randn([64, 64, 64, 64])
@@ -27,5 +27,3 @@ grad_xpu = copy.deepcopy(grad).bfloat16().to("xpu")
 output_xpu.backward(grad_xpu)
 print(input_xpu.grad.cpu())
 # print(input_xpu.grad.cpu())
-
-
